generated_apps/delivery-studio-preview/backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.6
pydantic>=2.9.2
jinja2>=3.1.4
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
email-validator>=2.2.0
